Separated out files from main for better organization
